+++ tests/examples/local-run.yaml
@@ -41,20 +41,20 @@ routing_table:
   inference:
   - provider_type: meta-reference
     config:
-      model: Llama3.1-8B-Instruct
+      model: Llama3.2-1B-Instruct
       quantization: null
       torch_seed: null
       max_seq_len: 4096
       max_batch_size: 1
-    routing_key: Llama3.1-8B-Instruct
-  - provider_type: meta-reference
-    config:
-      model: Llama-Guard-3-1B
-      quantization: null
-      torch_seed: null
-      max_seq_len: 4096
-      max_batch_size: 1
-    routing_key: Llama-Guard-3-1B
+    routing_key: Llama3.2-1B-Instruct
+  # - provider_type: meta-reference
+  #   config:
+  #     model: Llama-Guard-3-1B
+  #     quantization: null
+  #     torch_seed: null
+  #     max_seq_len: 4096
+  #     max_batch_size: 1
+  #   routing_key: Llama-Guard-3-1B
   # - provider_type: remote::tgi
   #   config:
   #     url: http://127.0.0.1:5009
